Remove CompareIgnoreCase
I noticed the giant lower_case_table[] for Mac when the formatting for it got
all screwy when clang-formatting.

But looking into where it was used, the only thing was for `gn clean` to not
remove "args.gn" when iterating the output directory. It seems like overkill
for that, so just ToLowerASCII() at that point.

Change-Id: I5673fbb22abfa70ae0ac4fb22c5a50899c06e686