Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> 
> How does this compare to the old buildtools binary's dependencies?
> 
> Also, I'm not wild about downloading something during a build script, but I think this is okay for now until we come up with something better. I'm not wild about adding a DEPS file for one hook, either.
> 
> Can you at least add a --no-sysroot flag to skip this for people that might not want it?

Yeah, I don't want to add gclient just for this. We could simply check the contents of the sysroot into third_party here too, if it's specifically the download that you don't like.

The other possibility is to do it in the bot recipe and let local users do "whatever", but I'd rather keep it always-on for now, so that there's less things to fix/debug at the moment.

Patch-set: 5
Reviewer: Gerrit User 1000004 <1000004@c360c8e5-7aa8-3905-a7a0-810934766dac>