Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> 
> How does this compare to the old buildtools binary's dependencies?

Unfortunately those look better:

 0x0000000000000001 (NEEDED)             Shared library: [libpthread.so.0]
 0x0000000000000001 (NEEDED)             Shared library: [librt.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [libm.so.6]
 0x0000000000000001 (NEEDED)             Shared library: [libgcc_s.so.1]
 0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]

in that there's no libatomic.so.1,and no ld-linux-x86-64.so.2.

So perhaps this will still be broken then, if it's the libatomic dependency that was causing the problem.

Unfortunately that's going to be non-trivial to resolve at this point, so I think we'd have to revert back to the in-tree binary.

> 
> Also, I'm not wild about downloading something during a build script, but I think this is okay for now until we come up with something better. I'm not wild about adding a DEPS file for one hook, either.
> 
> Can you at least add a --no-sysroot flag to skip this for people that might not want it?

Patch-set: 5
Reviewer: Gerrit User 1000004 <1000004@c360c8e5-7aa8-3905-a7a0-810934766dac>