Make gn --version work
Because GN_BUILD wasn't defined, gn --version was silently "UNKNOWN".
Implement a --version similar to the version that was used in Chromium
by using the number of commits on master + a fixed base larger than the
last commit position in Chromium. This will only really be meaningful
when built by the bots, but that's probably OK.

Change-Id: I744e2ae48cefd80eeb8d352ad67eda9daaac2385