[gn] Add export_compile_commands arg
Adding an arg to generate a compilation database (compile_commands.json)
in the out/ directory. Compilation databases are used by clang-based
tools (e.g. clang-tidy, clangd, clang-rename). CMake already does this
with the -DEXPORT_COMPILE_COMMANDS option, which regenerates the
compdb whenever the ninja files are regenerated.

Setting this flag to true adds ~900ms to the `gn gen` command.

`export_compile_commands` is set to false as a default here, and can
be overridden in the dotfile.

Change-Id: Ia3738e340b9de60da62557fb431a68a0618957fb