Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> GN should preserve the command-line flags across automatic regens (when Ninja re-runs it because a build file is out-of-date). If you run "gn gen <dir>" manually, they'll get lost, but I don't think people have occasion to do that for our project.
> 
> "fx set" will overwrite both the command-line flags and the build args so we'll have to teach it about this mode either way.
> 
> So I think the flag will persist as a command-line flag in a reasonable way for our build. If there's something significant I'm missing, we should fix it since I want people to be able to set these flags (I like "--color" myself).

I do run "gn gen" alone and want it not to lose state.  Moreover, we want people in Fuchsia to run "gn gen ..." without requiring any other flags for baseline normal behavior.  For Fuchsia, we want everybody to always generate the compdb so we can have scripts and such rely on it.
So for Fuchsia we would like to enable this in .gn or BUILDCONFIG.gn rather than on the command line.

Patch-set: 6