Update patch set 6
Patch Set 6:

> Patch Set 6:
> For Fuchsia, we want everybody to always generate the compdb so we can have scripts and such rely on it.
> So for Fuchsia we would like to enable this in .gn or BUILDCONFIG.gn rather than on the command line.

This is a new requirement to me. It's unfortunate this was not made more clear in the gn-dev thread. Can you please clarify this requirement on the thread? I think the code review is not the best place for this discussion.

If Julie wants to land this as-is and iterate, I think that's fine.

Patch-set: 6
Reviewer: Gerrit User 1000014 <1000014@c360c8e5-7aa8-3905-a7a0-810934766dac>