Support for Clang C++ modules
This implements the base support for Clang C++ modules in GN, see
https://clang.llvm.org/docs/Modules.html for background.

Rather than requiring developers to write module map files manually, GN
can generate these using the information inside the BUILD.gn files,
specifically the `public`, `sources` and `deps` variables. This
significantly reduces the effort necessary to roll out modules and
keeping the build files in sync with the source. This matches the module
used by Bazel.

Note that GN implements the explicit module compilation model where the
modules are compiled by invoking Clang with the appropriate flags, and
the compilation of individual translation units is dependent on the
module compilation step. This can significantly speed up the compilation
compared to the implicit model where the compiler manages the module
compilation automatically which leads to duplication of effort.

There are number of flags that control the compilation process, and the
open question is how many of these should be hardcoded inside GN vs
specified inside the build (see various TODO comments in this change).
In the latter case, we need to figure the right way to specify these
flags, i.e. how to express the fact that a particular flag should be
used for every target that uses modules. One potential option would be
to simply rely on configs, e.g.

  config("modules") {
    cflags = [ "-fmodules-strict-decluse", "-fno-implicit-modules" ]
    header_modules = true
  }

Note that `header_modules` variable isn't currently supported in
configs.

Change-Id: I7801dcd7e043c2c73af087e3457b785ca425aef4