Update patch set 1
Patch Set 1:

I'd rather not have to always use a command-line argument for this.
I was thinking of adding `check_system_targets` to .gn that's like `check_targets` but says matching targets also get `<...>` checking.  So `check_system_targets = [ "*" ]` would be the same as putting `--system` on the command line in your patch.

If you don't want to roll it into this change, that could be a follow-on feature and we'll get Julie to implement it.  She's the main person working on GN for Fuchsia now, so she's a good person to include on all reviews for changes motivated by our needs.

Patch-set: 1
CC: Gerrit User 1000023 <1000023@c360c8e5-7aa8-3905-a7a0-810934766dac>
CC: Gerrit User 1000002 <1000002@c360c8e5-7aa8-3905-a7a0-810934766dac>
CC: Gerrit User 1000016 <1000016@c360c8e5-7aa8-3905-a7a0-810934766dac>