Update patch set 2
Patch Set 2:

For what it's worth, we could really use this in our Vewd fork. We've so far avoided patching gn and rely on a horrible ABS_PATH method to achieve basically the same thing, which causes ugly side effects (absolute paths encoded in half the build, gn commands like 'path
 requiring full paths, etc.).

Frankly I don't think this is conceptually worse than the magic that toolchain variables are, or //build/secondary.

Patch-set: 2
CC: Gerrit User 1000047 <1000047@c360c8e5-7aa8-3905-a7a0-810934766dac>