Configure Xcode workspace to use old build system
The new Xcode build system cannot tie build error with a location
if path are relative, so create a WorkspaceSettings.xcsettings to
configure Xcode to use the old build system.

This is a workaround until the Xcode issue is fixed. Alternatives
are to pass -fdiagnostics-absolute-paths to clang (but this would
disable goma), use a wrapper script to invoke clang (increase the
complexity of the build) or use absolute path in generated ninja
files (goes against gn philosophy).

The behavior is controlled by --use-original-builder flag.

Bug: 852522
Change-Id: I257e0dda73fde413e33b4e2d2f36f5e50ea1b565