Extend the check_targets list format to allow black-lists.
Today the check_targets list is a white-list making it necessary
to change whenever new code is added in folders not generally listed.
It would be easier to maintain that list if it would just generally
list known broken targets.

To make that possible, add a "negated" label_pattern that will remove
a previous match, and make the last rule that applies determine whether
the filter list applies or not.

Bug: 896680
Change-Id: I3818deb3fb45e979e9205965f9c2dfc917a08d3c