Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Oh, so if we know what causes this problem, can you write more background on the CL description?
> 
> I feel like we shouldn't CHECK for user input problems, CHECKs are really for problems internal to the GN implementation. When it's more clear what is causing the problem let's see if we can find a different way to catch this issue.

The issue that triggered the ninja error is that in https://fuchsia.googlesource.com/zircon/+/sandbox/mcgrathr/gn-bug-16/public/gn/BUILDCONFIG.gn#304, the output_name for the static_library target is set to the _library_name. Then later, https://fuchsia.googlesource.com/zircon/+/sandbox/mcgrathr/gn-bug-16/public/gn/BUILDCONFIG.gn#354 declares a group named _library_name, which doesn't have the output_name specified and so defaults to the target_name as the name of the output file.

I'm not sure if this is the intended behavior, and I'm happy to go in and mess around with stampfile default names if that would be the preferred solution. If not, I can update this to print an input error message to the user earlier on (i.e. before the stamp file gets written).

Patch-set: 4
Reviewer: Gerrit User 1000023 <1000023@c360c8e5-7aa8-3905-a7a0-810934766dac>