Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Oh, so if we know what causes this problem, can you write more background on the CL description?
> > > 
> > > I feel like we shouldn't CHECK for user input problems, CHECKs are really for problems internal to the GN implementation. When it's more clear what is causing the problem let's see if we can find a different way to catch this issue.
> > 
> > The issue that triggered the ninja error is that in https://fuchsia.googlesource.com/zircon/+/sandbox/mcgrathr/gn-bug-16/public/gn/BUILDCONFIG.gn#304, the output_name for the static_library target is set to the _library_name. Then later, https://fuchsia.googlesource.com/zircon/+/sandbox/mcgrathr/gn-bug-16/public/gn/BUILDCONFIG.gn#354 declares a group named _library_name, which doesn't have the output_name specified and so defaults to the target_name as the name of the output file.
> > 
> > I'm not sure if this is the intended behavior, and I'm happy to go in and mess around with stampfile default names if that would be the preferred solution. If not, I can update this to print an input error message to the user earlier on (i.e. before the stamp file gets written).
> 
> @mcgrathr will likely have an opinion on this as well, I'm not positive what his goals are in that template.

I don't think the details of my templates are relevant.  stamp files are not something that exist as a concept in the GN language (aside from the bizarre requirement to define a "stamp" tool in every toolchain).  I should not need to know or care what the stamp files are called.  I think the simple way to achieve this is to derive the stamp file names from target names, since those already have to be unique.  But I don't care or want to know what they are called.  As a GN user, I should never have GN generate something that Ninja calls an error.  GN should generate valid Ninja, or it should diagnose something as an error in `gn gen`.

Patch-set: 4
Reviewer: Gerrit User 1000016 <1000016@c360c8e5-7aa8-3905-a7a0-810934766dac>