use a relative instead of absolute path mapping in the build dir
In order to refer to targets and sources outside the source root, GN
allows the use of absolute paths.  The location of the output files
and generated files from these targets include the absolute path to
the target, in order to avoid ambiguity.  But this makes reproducible
builds more difficult, and unnecessarily reduces cache hit rates for
build systems like goma.

In most cases where GN-based projects use absolute paths, the source
root is actually a submodule of a parent repository, so relative
paths would be the same between developers, and these would avoid
the problems mentioned above.  This patch makes GN use relative
paths with "UP" in place of ".." (and to represent an imaginary
directory above windows drive letters).

Change-Id: I827b0ad767dcf1e9b31df8bf7385dae9a70520f3