[metadata] Adding metadata output to gn desc
Converts the metadata scope to a base::DictionaryValue and outputs it as
the 'metadata' key in `gn desc`.

Also does a minor refactoring of desc so that single-value requests are
dispatched to the appropriate handler (no impact on performance or
output, other than allowing the metadata to be printed the same in both
single- and multi-value contexts).

Change-Id: If0beb25cdc0f05e9140e4768891bfbbac444ecc6