Improve help output.
Make the links in markdown for the "args" command and target variable
work by disambiguating href names in by variable type.

Special-case "gn help args" on the command line by printing help for
both the command and the variable.

Fix misnaming of xcode_test_application_name.

Simplify markdown tag emitting by requiring the the caller generate
a tag for linking rather than trying to autogenerate one based on
text that comes before a colon. This autogeneration code was not
used.

Change-Id: I7dd3f806eac5afaef746b2e37bf0903d9ae72b14