format: Fix incorrect dewrap of `if`
When `if (...stuff...) {` was exactly 81 characters, it would be
incorrectly wrapped because the trailing { wasn't being taken into
account in penalizing single line formatting.

To avoid this, pass the full suffix into the child formatter.

In a Chromium tree, diffing

  gn ls-files '*.gn' '*.gni' | xargs -l1 gn format

vs.

  gn ls-files '*.gn' '*.gni' | xargs -l1 /work/gn/out/gn format

results in
https://gist.github.com/sgraham/2e33ae36db6255af0be0785ea3d0b058

which are all 81 character long if statements being fixed.

Bug: gn:36

Change-Id: I5af9a62a3de4953ea0d60a85bd558281410a8f09