Format docs better for markdown
This edits the docs-generation code to put out docs in a nicer format
for markdown:

 * Don't wrap command line switches in a code block, let them render as
   a bulleted list.
 * Don't indent switches so much (otherwise they're in a code block
   anyway). 1 space is sufficient.

This has the unfortunate side-effect that command summaries (etc.)
have indent = 1 space (really, we want them to be 2nd-level bullets).
They still render in markdown as 2nd-level bullets (as desired), because
the adjacent 1st-level bullets have indent = 0 spaces. This is a bit of
a hack, but it avoids the complexity of passing "desired indent level"
to each invocation.

Test: out/gn help --markdown all > docs/reference.md
Change-Id: I8bfc547c23ed22fc0e172be2ec24238de77a00e8