Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> maybe? I don't see any docs in this module... how is it different than platform? can we merge them?

I need to add them, they're similar, the main difference is that while api.platform can only be used to detect host platform properties, this one can be used to construct arbitrary target which is what we need when cross-compiling.

If you're fine with merging it with platform, I'll do that a send a patch to recipes-py.

Patch-set: 2
Reviewer: Gerrit User 1000002 <1000002@c360c8e5-7aa8-3905-a7a0-810934766dac>