Update patch set 2
Patch Set 2:

I was thinking about waiting for a 2020 ping but might as well be proactive...

The old behavior was clearly wrong where you could write valid GN that produced in valid Ninja.

The current state would be that the developer is responsible for referencing the same pool, and if you want a pool to be global you have to declare everybody referencing the same pool in one toolchain.

This is not what people writing pools for linking want. But it's uncommon to write the linker rules so I'm not *super* concerned about this. Alternatives would be to issue an error if a pool is defined in multiple toolchains with the same name otherwise, or silently collapse them.

Part of the issue may be we added pools to actions. I don't recall exactly what the case for this was, but I think somebody was using pools for something more general than limiting memory for link steps, where having different pools for different toolchains may make more sense.

So I don't have a super strong opinion about how this should work.

Patch-set: 2