Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi, thanks for the CL.
> 
> I think we'd discussed this before and decided we should really have an arm bot doing continuous integration so we'd know if we broke something, before adding this code.

Ah, I see. Sorry for the duplicate. But these two files are not really third party code, they are part of Chromium's //base suite, so I'd assume they are pretty well tested and safe to use. If there is any doubt regarding this, it might be a good idea to get in touch with the team responsible for //base and ask them for their opinion and testing process.

Patch-set: 1
Reviewer: Gerrit User 1000119 <1000119@c360c8e5-7aa8-3905-a7a0-810934766dac>