Update patch set 12
Patch Set 12:

(1 comment)

Sorry for taking so long, I wanted to have a block of time to think about this.

I agree that this is theoretically correct. Also, I would encourage fixing mistakes in Chrome related to this regardless.

I have two concerns:

1. It's going to be hard to land. It will have to be coordinated with Chrome and Fuchsia. So Petr will also need to weigh in and we'll need to do a loosely coordinated landing so neither project is stuck on an unrollable version of GN tool long.

2. More generally, there's an assumption about compilation ordering following dependencies. I'm pretty sure the majority of GN users don't even know the difference between public_deps and deps. And it's hard for errors around this case to be caught and clearly attributed to the incorrect dependency specification, especially given the continued low usage of "gn check".

So in order to do this I would want a clear benefit. I don't see a discussion on build speed on either the CL description or in the bug. Do you know what effect this will have? If it's relatively small I think it would be better to keep the current behavior to avoid flaky builds. If it's large, then I agree it would be worthwhile despite the downsides.

Patch-set: 12
Reviewer: Gerrit User 1000014 <1000014@c360c8e5-7aa8-3905-a7a0-810934766dac>