Revert "Pass -lpthread when linking instead of -pthread"
This reverts commit 082fbe397d5217dc21b01c63c7f20eb6b77a3093.

That was the wrong fix; the compiler knows how to handle -pthread when
linking as well (in fact it even chooses the right position of -lpthread
when invoking the linker).

The original problem this commit was trying to fix was caused by the GN
bootstrap.py in the Chromium sources passing -nodefaultlibs to the linker
when the --use-custom-libcxx option is specified (which is the case for the
"Build From Tarball" bot), so -pthread ended up being ignored, and the
commit being reverted made -lpthread be explicitly passed to the linker
again.

This, however, breaks linking in other circumstances since we usually end up
passing -lpthread in the wrong position and the symbols are not resolved at
the right time. Use -pthread and leave it up to the compiler to figure that
out.

Bug: chromium:934793
Change-Id: Id4d4eeac12ed2b5b446a04e0ae681611feeaf66c