gen.py: Pass -pthread via ldflags, not libs
This reverts part of commit 0ac93dd2 ("For proper order put -lpthreads as a
'library' not a 'linker flag'").

As explained in https://gn-review.googlesource.com/c/gn/+/4020, we just need
to revert commit 082fbe39 and start passing -pthread in ldflags again, there
is no need to switch to using |libs| instead.

Bug: 934793
Change-Id: I28a32340f858dc8f5369840e6f7e6995d5f3e729