[DO NOT COMMIT UNTIL jumbo.gni FIXED] Also have gn check look at actions
Actions can also work on C/C++ type files in which case it
might be suitable to gn check them. This was one approach
to adding gn check support for jumbo but it doesn't currently work
since it's hard/not possible for the merge action to have
the correct dependencies because of visibility rules. It is
not allowed to depend on everything it needs to depend on.

Bug: gn:57
Change-Id: I8ee764cd525b430a528f9d72d1330c8066a9c046