Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I think having sources used by the build but never declared to the build system is problematic in a number of ways. For example, "gn analyze" will be totally broken and I don't think there's a path to make this work. I'm wary of adding complexity and exceptions to make one-off-things work. Feel free to post to gn-dev for more opinions.

I also think it's bad that the files can't be listed in sources, but what other solution is there? Would it be better to accept the native support for jumbo? It would be nice to unfork that gn fork.

From a C++ perspective there is nothing technically wrong with including cc files (and as you know it's common in unity build systems). This is also a small patch, the resulting error messages are good.

For gn analyze, that would also be more cleanly solved with the native jumbo support, though we have it working for jumbo through another hacky solution.

Patch-set: 3
Reviewer: Gerrit User 1000081 <1000081@c360c8e5-7aa8-3905-a7a0-810934766dac>