Have gn check handle multiple levels of included cc files
It is common in unity build systems, like Chromium's jumbo builds,
to have cc files include other cc files. Those are not checked
by gn check both because they are often generated, and because
gn check depends on all included files being listed in |sources|.

Included cc files can't be listed in |sources| because of the
side effects that would have so for those, have gn check follow
the include directly. This is limited to non-header files because
doing the same for header files would result in checking the
same files many times, possibly thousands of times, also
having to be able to handle include cycles. It is also not necessary.

This patch depends on patch that adds the --check-generated flag.

Bug: gn:57
Change-Id: Ia17e1ecd5a680513b53b51a0cd47539f1923edd6