Update patch set 1
Patch Set 1:

Oh, bit late, but can we get a test added for this? Normally I'm not picky but since this did actually break our tooling, it seems like a good idea.

Second, if I'm understanding this correctly, this means that runtime_deps didn't work for anything, right?

If so, I wonder why we only saw this on the skia bot, when presumably it should've broken everything in chromium, too. There may be something else to investigate on our side ...

Patch-set: 1