Move Tool logic out of Toolchain
This sets up the ability for tools to self-manage and dispatch, and
replaces the ToolType enum with the more dynamic tool name constant.

This is part of the Tool refactoring to allow for additional sets of
tools to be added, enabling support for new languages to be added. Same
as in the previous patch, this unfortunately requires touching a bunch
of files since the enum is prevalent.

Change-Id: Ia9eebb2525749aa0c525f85a9d89ce9ca6ca0159