Fix issues with bundle_data/create_bundle targets
Remove the bundle_plugins_dir variable as it is not really required
(the value is always the same for both iOS and macOS) and the value
can be inlined instead ({{bundle_contents_dir}}/Plugins).

Fix NinjaCreateBundleTargetWriter to only complain about missing
compile_xcassets tool if the target require it (i.e. if it defines
a list of resources to compile in an asset catalog).

Fix CreateBundleTargetGenerator to correctly report an error if a
required variable is not set and changed bundle_executable_dir to
not be required (not all bundle contains a binary).

Fix incorrect use of deps instead of public_deps in bundle_data
documentation.

Bug: 64
Change-Id: I28a0a3d673dc1b39804c466499e6e91f034a2113