Remove ToolType enum
Replaces it with the more dynamic tool name constant.

This is part of the Tool refactoring to allow for additional sets of
tools to be added, enabling support for new languages to be added. Same
as in the previous patch, this unfortunately requires touching a bunch
of files since the enum is prevalent.

Change-Id: I1d8f1b83a47196965f159ff9106c52a96f442f76