Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > Oh, bit late, but can we get a test added for this? Normally I'm not picky but since this did actually break our tooling, it seems like a good idea.
> 
> We don't currently have a setup for command tests (though if you want one, we can add abilities to redirect OutputString() calls to a particular ostream and refactor this command). It's nontrivial, though, and so if you wouldn't mind filing a bug for it?

Okay. I'm generally familiar with how easy (or hard) this can be given the way GN is set up, but I didn't look at the details of this in particular to see if it'd had gotten easier with your metadata changes.

I can accept "not easily" as an answer :)

> 
> > Second, if I'm understanding this correctly, this means that runtime_deps didn't work for anything, right?
> > 
> > If so, I wonder why we only saw this on the skia bot, when presumably it should've broken everything in chromium, too. There may be something else to investigate on our side ...
> 
> I'm less familiar with how Chrome/Skia uses runtime_deps, but yes, presumably it would have broken Chrome's mb script silently. Chromium likely hasn't rolled GN in a while.

Yeah, I want us (on the Chromium side) to understand what happened better here before we roll forward again.

Patch-set: 2