Let source_target_relative be a valid substitution for bundle_data.
This helps address the situation where one needs to package a bundle not
produced by the build's create_bundle rule, like a third-party
framework, inside another bundle that is produced by the build. If a
bundle_data rule only pointed to the outer directory of that third-party
framework, any changes to the inner components of the bundle would not
trigger a re-copy.

The {{source_target_relative}} is useful because it lets one define a
bundle_data target at the same level as the third-party bundle, list
its entire contents in |sources|, and have the bundle's folder
structure preserved.

With the existing set of valid substitution patterns, a bundle_data
rule would need to be created for each level of folder structure in
the third-party bundle.

Bug: 955936
Change-Id: I6e32a98ea607fb95d73ec707c7c8e9a39ee1e9d0