Move SourceFileTypeSet to Target
Have the target track which types of source files it contains, and check
that only one type exists. Prevents recompution in a few places, but
also allows for checking what language a target should be compiled
under.

Note that it will error out if complilation-incompatible source types
are used (e.g. having C and C++ in the same target is fine, but having C
and Rust in the same target isn't).

Change-Id: I5afb038e4cfa7da79c9f264cee79a8c652797dc3