Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+1
> 
> I think this CL is good (and I guess would allow fixing issue https://bugs.chromium.org/p/chromium/issues/detail?id=608382).
> 
> > Patch Set 2:
> > 
> > WDYT of this? I couldn't find other examples in the codebase of packaging a binary bundle into a bundle that is built by GN.
> 
> I think we do this all the time on iOS with the Plugins framework (they are framework containing an executable, see for example "search_widget_extension_bundle" target) and we do not need the new expansion for that. The following works:
> 
>   create_bundle("B") {
>     bundle_root_dir = "B.appex"
>     deps = [ ... ]
>   }
> 
>   bundle_data("B_bundle") {
>     sources = [ "$root_out_dir/B.appex" ]  // Note we list the app root dir
>     outputs = [ "{{bundle_plugins_dir}}/{{source_file_part}}" ]
>     public_deps = [ ":B" ]
>   }
> 
>   create_bundle("A") {
>     bundle_root_dir = "A.app"
>     deps = [ ":B_bundle" ]
>   }
> 
> This works because ":B" lists "$root_out_dir/B.appex" in its outputs and because there is a dependency between ":B" and ":B_bundle" thus if ":B" is rebuild, ":B_bundle" will also be rebuild.
> 
> Maybe I misunderstood what is the issue describe by the CL you've linked to. 

Right, on Mac we do that too and it works because "B" is produced by the build system. The issue I'm trying to address is that in the Keystone scenario, "B" is a prebuilt framework that is dropped into a separate repo and rolled via DEPS. I tried expressing this as:

bundle_data("B_bundle") {
  sources = [ "//third_party/B/B.framework" ]
  outputs = [ "{{bundle_root_dir}}/Frameworks/{{source_file_part}}" ]
}

But when the DEPS roll happens, git only updates the changed files, which are the inner components of the bundle and not the outer .framework folder. (Or, just `touch third_party/B/B.framework/B`). Because only the outer folder is listed as an input to the bundle_data rule, ninja doesn't know that the inner components need to be re-copied.

> 
> > The CLs that make use of this are:
> > https://chromium-review.googlesource.com/c/chromium/src/+/1603663
> > https://chrome-internal-review.googlesource.com/1282138
> 
> Still lgtm because I think this is a worthwhile change, but I'm not sure this is required for Keystone.

Does the above clarify? Can you think of a different way to do this (I tried several alternatives, including a copy() rule first, but because we implement that as links, that didn't work either)?

> BTW, can you add chromium:608382 to the Bug: tag in the CL?

Done.

Patch-set: 3
Reviewer: Gerrit User 1000166 <1000166@c360c8e5-7aa8-3905-a7a0-810934766dac>