Update patch set 3
Patch Set 3:

Is it okay if I revert this, or someone could look into the following fix ASAP? This assumes that all source files are created with the SourceFile(base::StringPiece) constructor, which is false -- most are actually created as:

SourceFile foo;
foo.value_ = ResolveRelative(...);

in various value extraction and file resolution functions across the codebase, and so this is broken for Chromium and for Fuchsia. Shouldn't be too hard to fix, hopefully.

Patch-set: 3
CC: Gerrit User 1000023 <1000023@c360c8e5-7aa8-3905-a7a0-810934766dac>