Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> (1 comment)
> 
> brettw ptal.  In Chromium, there's a couple hundred cases that need to be fixed, and I suspect Fuchsia will be similar.  It's mostly just mechanical removing of invalid source files.  Is it ok to break backwards compatibility like this?
> 
> This also fixes some bugs I introduced in my last CL where |type_| could get out-of-sync.

So as you note, there are a number of cases in Fuchsia where this is a problem, as in Chromium, and some of those breakages are in third_party code that is non-trivial to fix (e.g. protobuf). It also means that some relevant input files like .sym and .ipp files need to be added to the extension parsing, as they are reasonable to use in a buildable target, but that at least is just the fairly trivial matter of finding those and adding them.

This will be blocking a roll for us to start using and iterating on the Rust support that will be landing in the near future, as well as any bug fixes that get landed before this is resolved.

Would it be possible to notify the mailing list when breaking changes are being made in the future so that this isn't a surprise and make every effort to transition the breaking changes and/or provide sufficient time for other build maintainers to fix their own before landing changes like this? And, if the current issues can't be resolved relatively quickly, revert these changes until they can be?

Patch-set: 6