Update to more uses of C++11.
Adds move constructors to some fundamental types such as SourceFile,
SourceDir, and Label.

Removes most uses of swap and custom swap definitions in favor of move
semantics.

Converts most default class member initialization to declare default
values in the definition.

Moves more to using implicit and defaulted-in-the-header constructors
and destructors. The previous style was inherited from Chrome which had
those rules out of concern for code size. But for this small project the
header-based ones are a little easier to follow.

Change-Id: I7d5c3bd4ac9863c34b1072fdefef81e59e555a7f