Update patch set 6
Patch Set 6:

I figured out the problem. The default GN build is a release build which has assertions disabled, but the bot is running in debug mode. Do "build/gen.py -d" to convert to a debug build, the problem then occurs.

I suspect the problem is that AddGenDependency is expecting an absolute file path. If that's indeed the case, you should add a call to
  build_settings->GetFullPath(...)

Additionally (if that's the problem), can you add a comment in AddGenDependency in scheduler.h about this requirement?

Patch-set: 6
Reviewer: Gerrit User 1000015 <1000015@c360c8e5-7aa8-3905-a7a0-810934766dac>