Add command_launcher field to tools
This is a field to specify the command launcher for a given tool's
command. It appends the launcher to the command in the Ninja rule, but
keeps the two separate inside of GN. This allows the IDE writer tools
(e.g. the compilation command database) to be written without the
launcher prefix.

The presence of the launcher in the command has been breaking a number
of developer workflows in VSCode, CLion, and clangd. This should resolve
all of those workflows.

Change-Id: Idec6f530ddc48da01ac0c79b64032c7cc88856a4