Update patch set 1
Patch Set 1:

A number of issues have arisen recently to having to do with the presence of the goma prefix in the compilation database. With this, the compilation database generator would not include this field, and the Ninja rule writer would, which would unbreak a whole bunch of IDE support issues (e.g. VSCode, CLion, YouCompleteMe, clangd)

Patch-set: 1