Update patch set 8
Patch Set 8:

After thinking about this I would prefer not to take patches for this type of OS. We can do things like the main BSD flavors and AIX, but for more hobbyist-targeted systems I think it's better if you maintain your own fork if you need it. These things tend to bitrot rather quickly and normally there is not the continuous followup from the system folks required.

Patch-set: 8