Update patch set 10
Patch Set 10:

(1 comment)

> Patch Set 8:
> 
> After thinking about this I would prefer not to take patches for this type of OS. We can do things like the main BSD flavors and AIX, but for more hobbyist-targeted systems I think it's better if you maintain your own fork if you need it. These things tend to bitrot rather quickly and normally there is not the continuous followup from the system folks required.

I looked at the guidelines for a new port and I do not expect the Chromium team to maintain these patches or provide a builder for Haiku or for other non-mainstream OSes. Someone from the HaikuPorts team (myself included) will maintain the patches in this port of GN.

Since Haiku follows the POSIX standard, it already is compatible with most of the is_posix() OSes and is using OS_POSIX where ever possible, similarly with Mac, Linux and the BSDs so the platform-specific definitions are minimal. The only Haiku-specific code is the GetExePath() implementation and with this, Haiku passes all of GN's unit-tests. So adding Haiku as part of OS_POSIX makes sense here for less duplication and portability.

I find that upstreaming these changes here significantly reduces the patches maintained at HaikuPorts and allows Haiku to be a valid upstream target in other projects that use GN such as V8, Node.js and QtWebEngine, rather than maintaining platform support in our own fork(s).

Patch-set: 10
Reviewer: Gerrit User 1000155 <1000155@c360c8e5-7aa8-3905-a7a0-810934766dac>