Update patch set 2
Patch Set 2:

Thanks for writing this, it will be very helpful.

Currently all of the documentation is built-in which I kind of like. Is it possible to do that for this also?

At the bottom of "help" there's an "Other help topics" section. I would add this as one of those (think up a unique name for it).

Then we can also add cross-references from the metadata-related commands and variables to see "help <blah>" for the overview.

Patch-set: 2
CC: Gerrit User 1000014 <1000014@c360c8e5-7aa8-3905-a7a0-810934766dac>